requirements.txt
langgraph>=0.2.55
langchain>=0.3.15
langchain-core>=0.3.31
langchain-openai>=0.3.0
langchain-community>=0.3.15
langchain-experimental>=0.3.4
pydantic-settings>=2.7.0
python-dotenv>=1.0.1
chromadb>=0.5.23
scrapling>=0.4.0
httpx>=0.28.0
tavily-python>=0.5.0
langchain-tavily>=0.2.0